\name{Data}
\docType{data}
\alias{Data}

\title{Airline Passenger Dataset}

\description{
 Airline Passenger Dataset
}
\usage{data(Data)}
\format{
  \code{Data} is a data.frame with 156 rows and 1 column containing the yearly number of airline passengers
}

\keyword{time series forecasting}
\keyword{time series prediction}
\keyword{time series}
\keyword{sarimax}
\keyword{regression}
\keyword{linear regression}
\keyword{missing imputation}
\keyword{prediction intervals}

